Extract transaction amounts into separate module
